pyproject.toml
[build-system]
requires = ["setuptools>=60", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "linopy-yaml"
version = "0.0.1"
description = "YAML-based math definition layer for linopy"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
]
requires-python = ">=3.10"
dependencies = [
    "linopy>=0.4",
    "xarray>=2024.2.0",
    "pandas",
    "numpy",
    "pyparsing>=3.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
]

[tool.setuptools.packages.find]
include = ["linopy_yaml*"]

[tool.ruff.lint]
select = ["F", "E", "W", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
ignore_missing_imports = true
